requirements.txt
streamlit
Pillow
google-cloud-vision